[refactor][feat]차트 구현코드 리팩토링 및 데이터 operationId 기준 분리기능 추가(하드코딩 부분 차후 수정)
